use a signed size type when comparing

+++ main.cpp
@@ -466,7 +466,7 @@ bool safe_sendfile(int in_fd, int out_fd, size_t data_length) {
   __off64_t data_sent    = 0;
   ssize_t   data_written = 0;
   // Loop while there is data remaining and sendfile(...) succeeds
-  while (data_written >= 0 && data_sent < data_length)
+  while (data_written >= 0 && static_cast<size_t>(data_sent) < data_length)
     // Attempt to copy a chunk of data and record the amount written
     data_written = sendfile64(out_fd, in_fd, &data_sent,
       data_length - data_sent);
